Allow authenticated users without Minder projects to accept credentials
