+++ internal/util/cli/html/generic_failure.html
