+++ internal/util/cli/html/login_success.html
